if file exists in system than delete the file

+++ utils/utils.js
@@ -3,8 +3,12 @@ const fs = require("fs");
 
 const unlinkImage = (fileName) => {
   try {
+    const filePath = path.join(__dirname, "..");
     const file = path.basename(fileName);
-    fs.unlinkSync(`public/${file}`);
+
+    if (fs.existsSync(filePath + "/public/" + file)) {
+      fs.unlinkSync(`public/${file}`);
+    }
   } catch (error) {
     throw error;
   }
